Move CombatState into state-trackers directory
